requirements.txt
google-generativeai 
streamlit 
langchain 
python-dotenv
faiss-cpu
protobuf 
tiktoken 
langchain-google-genai 
pillow
pypdf 